Update patch set 1
No score.

> You have it backwards here - the phone is a dhcp -server-, and so
there
> is no address to be parsed until we get running. ESPECIALLY in the
> init.rc, which is - as far as I understand it - a fairly bizarre
little
> scripting language of it's own.
I've got it correct with phone as gateway.

udhcpd allows running of scripts when it writes leases file (actually it
occurs not once, but it's just example when it's possible to parse
leased IP, and get external IP of phone via ifconfig). As for init.rc -
you may write shell script and run it from init.rc, so syntax is not big
problem.

> This is an embedded device - indirection for indirection's sake is
> frowned upon. We're not going to have a bunch of endusers installing
> different core utilities in different places and generally making life
> interesting. The bins are set by the system, and when that happens the
> commands can be updated (once) and done.
Well, if official image contains no busybox, then it'll be impossible
use script with hardcoded path with this image. That was my concern.
Anyway, all my messages are just suggestions, your code seems normal to
me (it seems is doing what it is intended to do).

> Not in postrouting.
> # iptables -t nat -A POSTROUTING -i eth0 -o eth1 -j ACCEPT
> iptables v1.3.8: Can't use -i with POSTROUTING

I've searched a little bit  and found example of NAT for gprs connection
at OpenMoko, it's rather similar to what I've wrote:
 http://wiki.openmoko.org/wiki/Tethering

It's enough '-o' key only in POSTROUTING chain. '-i' is used in forward
chain related rule.

PS:
If you get this message twice or more times, '500 server error' is the
reason :)

Patch-set: 1