Update patch set 3
Patch Set 3: Fails; No score

I'm struggling to figure out how this change relates to https://review.source.android.com/12783 and I'm hitting merge conflicts in the Makefile.

Both seem to conditionally replace t32cb16blend.S conditionally on specific architectures (v7-a in one case, NEON in the other), and I'm not sure which is which.

If this change is still relevant, please upload a replacement with the minor makefile conflict take care of.

Patch-set: 3