Update patch set 3
Patch Set 3: No score

My preference would be that we bring mksh in as its own package (system/core/mksh), and have mksh and sh (ash) both look to TARGET_SHELL (or some similar build variable) to conditionalize the build.

if TARGET_SHELL is "mksh", mksh is built and installed as "sh".  ash is built and installed if TARGET_SHELL is "ash" or is undefined (to preserve the existing behavior). 

This allows the new shell to be enabled on a per-target basis, exercised, and once we are confident that everything is happy we can go ahead and nuke the old shell from the repository.

This allows us to, for example, decide to turn on mksh by default during gingerbread development so we can exercise the new shell, but be able to trivially switch back to ash if we run into any show-stopping issues.

Patch-set: 3
Reviewer: Brian Swetland <1001286@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0