Update patch set 1
Patch Set 1:

Hi Jean-Baptiste,
I think this change is still needed.  Problem is the kernel has also changed and so the futex() calls no longer work (and of course no one can be bothered to check syscall return statuses).  The kernel change was to make futex() require write access to the memory cell (imo, as a kludge work-around to a bug in futex).  To compensate, the mmap call in system_properties.c has to be changed to include PROT_WRITE.  So I should post that change separately as it impacts all uses of the system_properties.c module, not just watchprops.  Does that sound reasonable?
Thanks,
Mike

Patch-set: 1