Update patch set 1
Patch Set 1:

The concern is that this conflicts with Google's internal variant and I'm not planning to resolve the conflict since Google's current code solves both our and your problem.

I suggest merging your code in the open-source tree, and keeping Google's variant as the master copy, which will then overwrite your copy when it's released. At that point you'll be able to submit tweaks on top of Google's version as necessary.

Patch-set: 1