Update patch set 1
Patch Set 1: Verified

The code style looks fine, and the code merges and compiles fine into the open-source master tree.

However, the code that you modify was deleted in Google's internal master tree, and the direct consequence is that your change won't ever make it to an official release.

Reviewers, if you think that this is a good change, please +2 it, with the understanding that it'll only be in AOSP, and only until it gets replaced by Google's internal version of that code.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=+1