Update patch set 1
Patch Set 1: Fails

As you might have guessed, we don't normally accept platform contributions to gingerbread, as our processes make such contribution very unlikely to be accepted, and expensive to manage if they are. Contributions are normally sent to master, and it'd be great if you did so for this one (don't forget the remove the Change-Id to get a new one generated, so that the master variant doesn't collide with this one).

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1