Update patch set 3
Patch Set 3: Looks good to me, but someone else must approve

I fully agree with the fact that this patch is pretty limited as does not do the entire job by itself, but as explained in the readme.txt file because of Linux execve barrier I did not found any nice way that doing it this way.
What is it brings however is control of capabilities from init.rc file which is pretty flexible in case of generic services like mediaserver.
The counter patch will be posted on mediaserver ensuring it runs as "media" user (not the case today; not even enforced) + making use of this CapInh mask to set its capabilities.

Patch-set: 3