Update patch set 7
Patch Set 7: Abandoned

This is a semi-automated message. The details of your change have not been looked at as part of the process to send this message.

The last recorded activity on this change happened more than 10 months ago. That implies that this change was originally uploaded when Gingerbread (2.3) was the latest available version. Since Ice Cream Sandwich (4.0) was a very large update, it's very likely that this change doesn't apply on the latest source tree any longer.

Additionally, in August 2011 we had to turn off the servers that we used to be hosted on, and we had to restore from an older backup. In the process, we know that we lost about 7 weeks of data, and it's possible that the state as currently visible doesn't match the state that existed when those old servers were turned off, so it can't be trusted as is.

In those conditions, it's unfortunately unrealistic to investigate the current state of this change for inclusion in any future release. We are abandoning this change in order to focus on the more recent changes.

If you think that this change shouldn't have been abandoned, please cherry-pick it into a new topic branch, make sure that it looks correct and still compiles, amend the description to remove the existing Change-Id, and upload the new commit as a new change.

Thanks for understanding.

Patch-set: 7