Update patch set 1
Patch Set 1:

Not much we can do, that sounds like a bad implementation from the chipset vendor that conflicts with the way revision needs to work.  We could add a revision parameter to the command line, so at least future devices could pass a hardware and revision.  Mixing the hardware from the command line and the revision from the kernel doesn't make sense to me.

What do you need ro.revision for?

Patch-set: 1