Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

we really should remove all this code, it's totally legacy. We should never need to use it. The bootloader should put an image on the screen and the kernel should not reconfigure the display controller until userspace puts a new image on the screen (e.g. using the bootanimation, or recovery images, or charger images).

Patch-set: 1
Label: Code-Review=-1