Update patch set 1
Patch Set 1:

Thanks for reviewing. Many of the USB network adapters we are using don't seem to support Wake-on-LAN but this may be due to a mis-configuration on our part. I'll abandon this for now and we'll look into that approach further.

Patch-set: 1
Reviewer: Andrew P Boie <1011333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0