Update patch set 5
Patch Set 5:

Ok, I guess I'll abandon this again unless others want to chime in on it.  For our purposes, it was to allow setting ownership of selinuxfs booleans (/selinux/booleans directory) to permit writing by system_server and/or system UID apps ala SEAndroidManager without having to maintain a separate list of the booleans in the init.rc.  Set of booleans defined by policy configuration, not hardcoded anywhere.

Patch-set: 5