Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Your initial use-case was for ownership changes. And the complaint was that sysfs was not supported in ueventd.

Change your commit message.

Add support in ueventd, not in init who's readme predates the uevent*rc support.

Have the uevent rule express the fact that it needs to do an init trigger. Probably have it pass in some ID that could be common to multiple rules that care. That ID would then be passed to the normal init triggering mechanism "<name>=<value>"
The rule could end in something similar to:
   <normal rule> trigger:my_id
which would cause a "setprop uevent_my_id={add,remove,...}"

This makes it easier to support multiple devices nodes (/dev/ and /sys/) in one shot without the partial hack and substring matching which doesn't work for sysfs where the non-deterministic portion is in the middle of the path (see the example in my first comment).

Patch-set: 1