Update patch set 1
Patch Set 1:

JP... I really don't agree with where you are going with this. This patch is IMO really very clean and implements functionality that is well specified in the init README.

Yes... ueventd can/should be used for simple permission settings on device insertion/creation... but there are sometimes more complex cases. I don't think that extending ueventd.rc is the right solution... that is just going to eventually get out of hand.

I agree that the commit message should be revised. The message really shouldn't imply that this is a method that should be used for permission setting... instead the commit message should indicate that in certain cases, more arbitrary actions are required on device insertion/removal and this is the recommended method of triggering those actions.

... Can you live with that?

Patch-set: 1
Label: Code-Review=0