Update patch set 1
Patch Set 1:

Feel free to turn on -Wno-unused-parameter in your own projects, but that is not something we are going to put into every other project by default. IMO, unused parameters can signal bugs, so I feel that you should only flag the ones that you know have a reason for being unused and leave the warning enabled.

Patch-set: 1