Update patch set 1
Patch Set 1:

Sure I can do that. These sorts of warnings happen in all sorts of projects though. We're not currently explicitly focusing on fixing all such warnings but if there's only a few warnings in a project we tend to fix them to clean up the build. Especially if -Werror is in use. Should we create the macro for every project? Or can we create the macro in a system/core header somewhere that can be included in every project?

Patch-set: 1
Reviewer: Edwin Vane <1010717@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0