Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This commit has already been discussed on android-review.googlesource.com in another upload. We have decided to re-code it, taking into account the comments on the other review.

Thank you for catching this. It's an escape... we didn't intend to upload it again.

Xiaokang... please abandon this one.

Patch-set: 1
Label: Code-Review=-1