Update patch set 1
Patch Set 1:

I advised Andrew to remove this functionality because noone really should be using it, and on our side the code has rotted. It's not being maintained at all.

To your concern, the logo code is not the answer. In your scenario, the framebuffer is going to be handed over to linux anyway, so you are going to see garbage on the screen. With the logo approach, you will see it for a shorter period of time, but you'll still see it which is a poor user experience. Init shouldn't be touching the framebuffer, it should hurry up and get out of the way so that surfaceflinger can put the bootanimation on the screen.

The way to do this without a glitch is to have the bootloader pass the framebuffer information to the kernel where you can memblock_reserve it, pass it to your fb driver, and point the framebuffer at it.

Patch-set: 1
Label: Code-Review=0