Update patch set 4
Patch Set 4:

The #if 0 has been there since the code was first written:

http://s9/?change_num=57124

It seems like this change should be OK, but who knows if it might break some strange case I can't think of, and it does change the semantics to be more like an rsync instead of a cp. Is it really a big win do do this?

Patch-set: 4
Label: Code-Review=0