Update patch set 2
Patch Set 2:

What is the intended use of these devices?
For instance what does it mean when you select AUDIO_DEVICE_OUT_FM_TX on an output stream?
Or is it the indication of a particular input source to the audio subsystem? In this case how do you specify use cases like FM radio to Speaker or FM radio to Headset?
I suppose the input device is to capture from FM input for DVR type application?

Patch-set: 2