Update patch set 1
Patch Set 1:

We are trying to extend Android so that a single binary image works on multiple devices (we have lots of different development boards floating around internally which we'd all like to have Android working at some baseline level of functionality). In order to support this, we are trying to migrate some decisions which currently must be made at build time, to either be auto-detected at runtime or specified at installation time.

So to support install-time decisions, the installation/provisioning tool creates a special small partition where installation parameters are saved in a property file. This additional command is so that we can tell init to read it. We didn't want to put these properties on /data since they shouldn't be erased, nor have the installer modify /system since that needs to be fixed at build time to support incremental OTA updates.

Patch-set: 1