Update patch set 1
Patch Set 1:

I don't like changing the links to hard links.  Maybe have init print an error message to klog when chown/chmod is used on a symlink, instead of silently ignoring it?  Right now we silently ignore lots of errors, maybe we should go through and make a few of the ones that shouldn't happen in a sensible init.rc more noisy.

Patch-set: 1