Update patch set 1
Patch Set 1:

I see, you're trying to take advantage of the by-name linking.
The only reason I like the symlinking is it makes it easy to determine which block device and partition is which with ls -l /dev/block/platform/*/by-name/.  This sounds like a legitimate use of chowning symlinks, and switching to hardlinks is just a workaround for (IMHO) an overly restrictive security policy.

A few other things I'd like to explore before accepting this:
Relax the symlink restrictions on chown or introduce a new chown-equivalent that follows symlinks, as long as all of the directories above the source and destination are only writable by root
or
Put the real locations of the partitions into properties, so  you can do something like chown system ${ro.partition.platform.your_device.misc}

Patch-set: 1
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0