Update patch set 1
Patch Set 1:

I agree with Colin that this is a workaround and we should really address the underlying problem.

Putting everything into properties sounds like overkill to me and would be a last resort.

Firstly, we shouldn't be doing the chmod/chown on the device nodes in init when ueventd is the one that created the dev files and set its initial ownership/permissions. We should be able to specify everything in ueventd.rc somehow. Don't have any immediate solutions, but let me think on this, unless someone has some ideas.

Patch-set: 1
Reviewer: Dima Zavin <1000413@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0