Update patch set 1
Patch Set 1:

Instead of doing this, we could have made our bug reporting app AID_SYSTEM, but that would be an unnecessary privilege escalation.  It doesn't need to do most things that AID_SYSTEM can do, and we don't want it to be able to do so.  We want to selectively open up debugging capability to a signedOrSignature platform app without significantly increasing the risk of a security breach.  Using AID_LOG seems like a natural fit, because it's a system UID with limited capabilities.

I could probably detail this justification better in the commit message if that's your preference.

Patch-set: 1