Update patch set 1
Patch Set 1:

That would work except for the privileges that are not available to Java apps.  I think that would be CAP_SYSLOG at a minimum.  I'd have to look at which of the GIDs are available and necessary.

This also doesn't help with the original stated case, which is allowing the app to start/stop a bugreport.  It's unfortunate that allowing start/stop also allows it to direct where the data goes (thus allowing read access).

On existing Nexus devices, the bugreports triggered via the keypress are dumped into the sdcard, and therefore readable by even 3rd-party apps, but perhaps that's being patched up on your end.

Given that our goal is to debug system issues holistically on controlled sets of devices, perhaps we really should just use AID_SYSTEM.

Patch-set: 1
Reviewer: John Michelau <1005146@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0