Update patch set 10
Patch Set 10:

As a style thing, two exit points is pretty clear: a successful exit and an error exit.  Especially when the successful exit is reached without gotos and explicitly returns a good exit code ("return 0;") and the error exit, which is jumped to with a goto, cleans up any intermediate state, and returns with it's error code ("return rc;").

The readability of this (clear success vs clear failure) is helped by having two exit points.  To me this is much clearer than having a backwards goto.

The code as-is can be cleanly adapted by just having a second (failure) exit return instead of the backwards goto.

I know we don't really have a published C style guide, but I guess that's what these reviews are used for :)

The using negative error codes as return values (instead of just returning -1 and setting errno) is another stylistic thing.

Other style things are using 4-space indentation, but that's trivial to fix.

Patch-set: 10
Label: Code-Review=0