Update patch set 1
Patch Set 1:

This is only a first version. I haven't made a lot of testing (I intend to use this to run the full Chrome test suite through it), though I'm confident it's working correctly.

On the other hand, it requires a newer adbd running on the device, which makes its usefulness limited. At a minimum, I'd like to improve error reporting when trying to run the command against a "legacy" device (for now, all you get is "error: closed", which is not very meaningful).

I didn't find an easy way to test for features / version of the adbd daemon that runs on the device from the host, this might make this easier.

Also, I'd like to add an alternative solution that would work with legacy devices. For Chrome, we're using a set of two daemon processes (one launched on the device, the other on the host), who perform a complicated dance through adb forwarded sockets. It's fairly complicated to setup though.
I'm aiming at something simpler, for example:

1/ Have the developer push once a helper program into /data/local/tmp/, e.g.

  abd push adb_reverse_proxy /data/local/tmp/

2/ Have the "adb reverse" implementation on the host ADB server detect a legacy device, launch a new instance daemonized instance of adb_reverse_proxy through "adb shell" or equivalent.

I'm not sure of all the security consequences (the user could install any other program to the same hard-coded location), what are the chances that this would constitute a possible security attack vector (each new daemon would be started with "adb shell", I don't see an obvious escalation issue, but still).

I'd be interested in Nick's opinion on this. I can probably provide a more detailed design doc for this though.

For now, your input on the current patch would be appreciated :)

Patch-set: 1