Update patch set 1
Patch Set 1:

As for not resetting the supplemental group: part of this is intentional.  An application may not have the INTERNET permission (group inet), which prevents it from opening up sockets.  gdb, in particular, needs to open sockets for debugging.  If we don't give the app, via run-as, group inet, then debugging a non-INTERNET permissioned app becomes impossible.

Patch-set: 1