Update patch set 1
Patch Set 1:

Sorry, I thought I'd included the warning message in the initial comment here.

I believe the warning (in GCC) was:
 Warning: comma at end of enumerator list
or something like that. Pretty sure it's part of -pedantic.

I agree it makes prettier diffs to have a comma there; but it's apparently not valid C++ according to the 89 standard, per the first answer at http://stackoverflow.com/questions/1002462/gcc-is-it-possible-to-disable-the-comma-at-end-of-enumerator-list-warning-whe

Also: I'm actually not seeing this warning in Firefox-for-Android builds anymore, so it's totally possible that someone came along and changed the compile options such that GCC no longer spams this warning.  So I'm happy to consider this WORKSFORME (patch no longer needed).

Patch-set: 1
Reviewer: Daniel Holbert <1017373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0