Create change
Patch-set: 1
Change-id: Ia935dfc3fbcf61bb0b7a66813a8561c9ab04b1f8
Subject: Add a module to operate on services managed by the init daemon. This code is duplicated in a few places in the Android codebase so this change is the beginning of a consolidation effort.
Branch: refs/heads/master
Commit: 17aad23531c3f2d5ecad7079fed6d6bda1a6b152