Update patch set 1
Patch Set 1:

The reason I put them under the name of the disk controller is that I was worried about finding the wrong partition if an SD card or usb storage was connected that had a GPT partition table.  It would be trivial to root a device if it accidentally mounted an external partition as /system.

Patch-set: 1
Label: Code-Review=0