Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Add a module to operate on services managed by the init daemon. This code is duplicated in a few places in the Android codebase so this change is the beginning of a consolidation effort.
Commit: 75f4532e33f6e6c0061e4a0e12d0810bb82179f5