Update patch set 2
Patch Set 2: Code-Review+1

Adding extra options to the kernel command line through a default kernel config still need to get parsed by the init program.  So that isn't a work around.

We append an assortment of data items to the command line that is used in kernel mode and with this constraint we need to be sure that all the data passed to user mode through the command line (such as usb serial number) is put in the first 512 bytes.  

Basically we are at a limit where if we want to add pxe-boot NFS root options to the boot command line we are in trouble if we don't make sure the androidboot.* var's are not in the lower 512 bytes.

Its not a blocker but it is inconvenient to sort the cmd line for android platform data hand offs.


An example of a command line we create:
console=ttyS0,115200 console=logk0 earlyprintk=nologger loglevel=8 drm.debug=0x0 kmemleak=off ptrace.ptrace_can_access=1 emmc_ipanic.ipanic_part_number=3 androidboot.bootmedia=sdcard androidboot.hardware=byt_t_ffrd8 androidboot.spid=xxxx:xxxx:xxxx:xxxx:xxxx:xxxx androidboot.serialno=01234567890123456789012345678901 oops=panic panic=40 nmi_watchdog=panic softlockup_panic=1 vmalloc=172M vga=current i915.modeset=1 drm.vblankoffdelay=1 acpi_backlight=vendor i915.i915_rotation=1

The spid thing is an intel specific addition to communicate some FRU information to some of the intel HAL components.  This cmd line is 497 long and if we add nfsroot for pxe we go over the limit.

Patch-set: 2
Label: Code-Review=+1