Update patch set 3
Patch Set 3:

Dima,

I put the init change in a separate patch as requested.

The reason why we are stuffing so much stuff into BOARD_KERNEL_CMDLINE is because we are trying to support a large number of different boards with a single bzImage. In order to do that we can't hard-code a lot of stuff in the kconfig as the kconfig is common to all of them.

I realize this is a different approach from the typical way of having a different kconfig for each TAREGT_PRODUCT, but it simplifies things for us greatly if we can just drop any kernel into a build without worrying about target-specific options when the bzimage is generated.

Patch-set: 3
Reviewer: Andrew P Boie <1011333@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0