Update patch set 3
Patch Set 3:

The bootloader isn't device-specific either. It does set androidboot.serialno but the rest of the stuff is set in the build via BOARD_KERNEL_CMDLINE. We have one bootloader for EFI devices and another for IAFW.

Are you saying that you want the bootloader built with various extra command line information for every supported product and then append this info before starting the kernel?

This seems a lot more complex than simply having this stuff set on a per-product basis in BoardConfig.mk.

Patch-set: 3