Update patch set 1
Patch Set 1: Code-Review-2

We really do not do this in init, and I prefer that you didn't. I agree that putting it on the end is not ideal, but adding a -R like isn't either. Init.rc files aren't shell, so commandline compatibility is not relevant.

If you want a recursive option, I'd add a separate command restorecon_recurse or maybe restorecon_dirs or something like that.

Patch-set: 1
Label: Code-Review=-2