Update patch set 5
Patch Set 5:

Yes, I had a chat with Kenny to discuss this: the main security issue is that when receiving a packet, the code in adb.c:handle_packet() that deals with A_WRTE never checks that the remote id (i.e. p->msg.arg0) corresponds to an opened communication channel.

This means that a device A running a compromised adbd could spoof messages from another device B, if it can "guess" the corresponding host local id for an existing channel between B and the host.

I can provide a separate patch to fix this, but the current one isn't impacted by this.

Patch-set: 5
Label: Verified=0