Update patch set 6
Patch Set 6:

You can't determine that the log entries are unnecessary, or that logwrapper is a bad memory tradeoff for what the OEM is using it for.  I have no problem removing specific instances of logwrapper, but I disagree with removing it wholesale.

As for SELinux, that sounds like a problem that needs to be solved in SELinux, and not by deleting logwrapper.  Can logwrapper determine the context of the executable it wants to run and then drop to that context?

Patch-set: 6