Update patch set 1
Abandoned

This is not necessary. I should've abandoned it earlier.

(The context is that FileMap does strdup the string for us, but I initially diagnosed the backtrace in 12173498#1 incorrectly. malloc_debug was failing because we were freeing an invalid pointer, and not because we were freeing it twice)

Patch-set: 1
Status: abandoned