Update patch set 2
Patch Set 2: Code-Review-1

This change appears to be defeating the inlining of atomic routines and as such would be a performance regression (hence the -1). Long-term plan is that these atomic routines should disappear:
1) there are better ones in bionic and we shouldn't be maintaining 2
2) compiler intrinsics are better
We'd ideally use std::atomic but stlport lacks them and we've not switched to libstdc++. I believe to clean these things up in the shortest amount of time we should stop writing inline assembly and switch the implementation for all architectures to use __sync_bool_compare_and_swap and similar. Is this compatible with clang? Thanks!

Patch-set: 2
Reviewer: Ian Rogers <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1