Update patch set 2
Patch Set 2:

logwrapper should already transition domain upon executing the daemon, by virtue of the fact that it will stay in init's domain by default and there is already a domain transition to the daemon domain on executing the daemon executable.
The problem that remains however is that any sockets pre-created by init for the daemon (based on socket entries in the init.rc) will be wrongly labeled without this change.  We did upstream a change to support specifying the security context of the socket, so that is possible instead but it requires explicit specification of the context in the init*.rc service entry rather than automatic computation from policy.  We were using that in one of the device trees at some point, probably mako, until the logwrapper uses were dropped.

Patch-set: 2
Reviewer: Stephen Smalley <1010111@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0