Update patch set 1
Patch Set 1:

Yes, we can take this. I agree with the other Chris that the mips code wasn't looking at eh_frame. I think I can fix that relatively easily, or at least explain how to fix it. The only other problem was that the user context passed for the thread unwind doesn't include a valid stack pointer. I assume that's actually a kernel bug, but is also present for libcorkscrew.

Patch-set: 1