Update patch set 1
Patch Set 1: Code-Review-1

atomics are, for the vast majority of cases, used for operations on int32_t, or synonyms of. Changing the behaviour of the existing functions to operate on "longs" is inappropriate as on 64 bit platforms, the majority of those calls will still operate on 32bit types.
The consequence of this change is that many projects will need to be changed to use the "32" version of the functions to operate as they did before. 
I don't believe this is advisable, and that we should explicitly add 64 bit atomics - operations that 32bit platforms can also implement and call.

Where your approach makes more sense is when handling pointers. There we can make a uniform decision to use "atomic pointer" operations that work on the correct size. Counters tend to be a fixed size, int32_t.

Patch-set: 1
Reviewer: Stuart Monteith <1037216@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1