Update patch set 1
Patch Set 1:

I guess it depends on your goal; if you just want to be able to make shell enforcing in all builds (but leave adbd itself permissive in userdebug/eng), then this patch suffices.  If you want to be able to make shell and adbd enforcing, then you'd need to setcon() or have init restart adbd in a different domain for the adb root case.

Patch-set: 1
Label: Code-Review=0