Update patch set 1
Patch Set 1:

Rather than put the context into code, we could:
a) pass it as an argument from init.rc, so that it appears in the init.rc just like the seclabel entry,
b) compute it in the same way as is done for service contexts by init if we treat it as a simulated exec of su, i.e. getfilecon of su, call security_compute_create(), and setcon() to the resultant context, or
c) have adbd read it from a separate config file in the rootfs.

Patch-set: 1