Update patch set 1
Patch Set 1: Code-Review-1

This is too simplistic of a solution, and would inhibit the use of restorecon_recursive in the future should we find a need to change the labeling again.

Instead, I'd prefer if we create a one-shot service, run from init, which reads a configuration file from the root partition and compares it to a configuration file on /data. If there are fixup steps which exist in / but don't exist in /data, then we should perform the fixups. Afterwards, the file on /data would be updated to indicate that all the fixups were performed.

Patch-set: 1
Label: Code-Review=-1