Update patch set 1
Patch Set 1:

With regard to preventing future restorecons, we can always unlink the .restorecon_done file if we know we need to relabel that subtree again, so this approach does not preclude that.

With regard to any service-based approach, I'm not sure how to ensure that it runs before any other service that might try to access the relevant directories.  Doing it directly as part of the existing restorecon_recursive at the same point where the directory is created and chown/chmod'd seems safest in that respect.

Patch-set: 1
Label: Verified=0