Update patch set 2
Patch Set 2:

Colin / Stephen: Another option is similar to what Stephen suggested, and creating a new builtin which implemented my initial suggestion. Something like:

restorecon_fixup: Read /sefixups and compare it to /data/sefixups. If there are entries in /sefixups which aren't in /data/sefixups, perform the necessary fixups, and update /data/sefixups to indicate success.

That way, all of our restorecons would become one-shot, and we wouldn't take the performance penalty on every boot. 

This *might* even allow us to relabel all of /data if we could guarantee it would only occur on the first boot after an OTA.

Patch-set: 2