Update patch set 2
Patch Set 2:

In general we don't want init to block on anything.  ueventd is the only exception - it used to be part of the init process, but due to a deadlock between module loading and responding to kernel firwmare requests we had to make it a separate process (we also don't want threads in init).

exec has never been implemented for the same reason - we don't want blocking, and it will immediately get abused in ways that reduce overall system stability.

I'm hesitant to block init on something as complex as sefixups, and on something that depends on /data - what happens if /data is corrupted?

If we do end up with something like this, I suggest putting it inside fs_mgr, which already blocks init for filesystem mounting and fscking.  Treat it as part of the /data mounting process.

Patch-set: 2