Update patch set 2
Patch Set 2:

I don't like scattering these files around the filesystem, especially without a clear understanding of exactly what they mean.  It's only a matter of time until we have to relabel and already-labelled directory, and then we're back where we started - there is no way to know if the directory is labelled correctly or not.  Sure, you can remove .restorecon_done, but /data is not mounted during an ota, so who is going to remove it, and how do they make sure they only remove it once?

Patch-set: 2