Update patch set 2
Patch Set 2:

Hmm...well, one option would be to check the context of the top-level directory against the expected value and perform the restorecon_recursive if they do not match even if the .restorecon_done file is there.  Then if you ever change the context for that directory, it will fix it on the next boot.

Patch-set: 2