Update patch set 2
Patch Set 2:

Skipping restorecon_recursive if the context on the top-level directory is already correct will work for the persistent directories (/data, /persist), but not for sysfs.  I guess if we go that route we want a separate restorecon_recursive_conditional built-in or similar to distinguish whether we want it applied conditionally or unconditionally.
Will try to look at fs_mgr but not really clear on that route.

Patch-set: 2