Update patch set 1
Patch Set 1: Code-Review+1

I like this idea.

The path to the ctrl_interface sockets are always passed in to the wpa_supplicant service as an option (-O), which is why I believe the various mkdirs are in the device specific init*.rc files to begin with. By moving that directory creation out to  core we could potentially force future device makers to use the /data/misc/wifi location which is already correctly labeled for them.

Patch-set: 1
Reviewer: Robert Craig <1013887@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1