Update patch set 3
Patch Set 3:

Colin,
I do catch your point we have one additional patch to track LOCAL_*FLAGS/INCLUDES variables involved at build time which change between subsequent builds:
https://android-review.googlesource.com/#/c/87557/

That will automatically rebuild modules when required, but that breaks toolbox precisely because the source files depend on a generated header (tools.h) which is not regenerated between builds even when ALL_TOOLS variable changes.

But this is the only module which breaks, no other module suffer from the same issue.

Your suggested solution would work but would not be as granular as this one, it would induce the rebuild of the full module and not just the dependent sources.

Patch-set: 3