Update patch set 1
Patch Set 1:

This is where I need some wisdom from both of you. DO NOT MERGE is me being a coward.

I want to turn on the floodgates of more independent testing, but I do *not* want to affect current programs until I have a plan for the kernels. There is an inherent inefficiency once we flip this switch resulting in both the kernel and the user space taking up resources. Propagating to any program that has a .config with CONFIG_ANDROID_LOGGER in the associated kernel is IMHO a bad thing. In addition, the gTest that checks for this duplication, *and* any duplicate leakage, of resources is currently blocked behind review, and we do not have a plan for CTS or Q/A testing. I figure AOSP is the only place that can take this bruise right now. Finally, there has been no independent svelte review of the user space logger (likely a formality).

This change, given all the above, is premature.

Patch-set: 1
Label: Verified=0