Update patch set 1
Patch Set 1: -Code-Review

I was deeply concerned with the recursion that would result in a KISS implementation. I can address your concern by separating the code that initializes the logd_fd from the implied call to send the message. This will actually work better as I was also concerned about multi-threading and the atomic update of write_to_log ...

Patch-set: 1